+++ sign_recognizer/model/utils.py
