+++ sign_recognizer/parsermedia/camera.py
